pki (mTLS) cert load testing only
This is a personal fork that focuses only on load-testing the PKI
backend. All of the other loadtest such as the KV backend are disabled.
This also means you can avoid some of the setup around seeding initial
secrets.

There is also some code that can be uncommented/commented if you want to
test a certificate role with generate_lease=false/no_store=true